requirements.txt
matplotlib==3.5.1
numpy==1.22.2
pandas==1.5.3
python-whois
wayback